Add files via upload
It's an implementation for converting infix expressions to postfix notation. The code uses the Shunting Yard algorithm and includes functions for tokenizing the expression and performing the conversion. It's a handy tool for evaluating mathematical expressions or implementing algorithms that require postfix notation. Take a look and let me know what you think! #C++ #Coding #Algorithm"
 #calcolator
#projectcpp